feat(desktop): polish agent activity feed rows
- Replace the monolithic AgentSessionToolItem implementation with a barrel export and focused render modules for the main tool item, compact rows, sent-message context, shell output, todo summaries, image previews, detail blocks, and message-link helpers.
- Render shell command activity details as a muted terminal-style block below the existing accordion label, with inline accent-colored terminal icon, muted command text, and foreground stdout output.
- Parse shell tool result envelopes so stdout displays without exposing the raw JSON result payload, and add focused parser coverage for envelope and raw text cases.
- Keep plan update progress summaries, such as 5/5 complete, in foreground while preserving muted surrounding label text.
- Render todo tool summaries as static checkbox snapshots in a dedicated component, keeping the desktop file-size guard green.
