+++ desktop/src/features/agents/ui/AgentSessionToolItem.tsx
@@ -1,812 +1,1 @@
-import * as React from "react";
-import * as DialogPrimitive from "@radix-ui/react-dialog";
-import { CheckCheck, ChevronDown } from "lucide-react";
-
-import {
-  resolveUserLabel,
-  type UserProfileLookup,
-} from "@/features/profile/lib/identity";
-import { cn } from "@/shared/lib/cn";
-import { rewriteRelayUrl } from "@/shared/lib/mediaUrl";
-import { normalizePubkey } from "@/shared/lib/pubkey";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/shared/ui/dialog";
-import { UserAvatar } from "@/shared/ui/UserAvatar";
-import type { AgentActivityAction, TranscriptItem } from "./agentSessionTypes";
-import { getBuzzToolInfo } from "./agentSessionToolCatalog";
-import {
-  buildCompactToolSummary,
-  type CompactFileEditSummary,
-} from "./agentSessionToolSummary";
-import type { FileEditDiff } from "./agentSessionFileEditDiff";
-import { FileEditDiffBlock, hasFileEditLineDiff } from "./FileEditDiffView";
-import {
-  ActivityRowLabel,
-  splitActivityRowLabel,
-  type ActivityRowLabelParts,
-} from "./activityRenderClasses/ActivityRow";
-import { TranscriptTimestamp } from "./activityRenderClasses/TranscriptTimestamp";
-import type { AgentTranscriptIdentityProps } from "./activityRenderClasses/types";
-import {
-  asRecord,
-  formatCodeValue,
-  formatTranscriptTimestampTitle,
-  getToolString,
-  getToolDurationDisplay,
-  isInlineImageData,
-  parseToolResultValue,
-} from "./agentSessionUtils";
-
-export function ToolItem({
-  agentAvatarUrl,
-  agentName,
-  agentPubkey,
-  item,
-  profiles,
-}: AgentTranscriptIdentityProps & {
-  item: Extract<TranscriptItem, { type: "tool" }>;
-  profiles?: UserProfileLookup;
-}) {
-  const [isExpanded, setIsExpanded] = React.useState(false);
-  const hasArgs = Object.keys(item.args).length > 0;
-  const hasResult = item.result.trim().length > 0;
-  const canonicalToolName = item.buzzToolName ?? item.toolName;
-  const buzzTool = getBuzzToolInfo(canonicalToolName);
-  const compactSummary = buildCompactToolSummary(item);
-  const duration = getToolDurationDisplay(item);
-  const messageLink = getSentMessageLink(item);
-  const timestampTitle = formatTranscriptTimestampTitle(item.timestamp);
-  const agentProfile = profiles?.[normalizePubkey(agentPubkey)] ?? null;
-  const agentLabel = resolveUserLabel({
-    pubkey: agentPubkey,
-    fallbackName: agentName,
-    profiles,
-    preferResolvedSelfLabel: true,
-  });
-  const agentResolvedAvatarUrl = agentProfile?.avatarUrl ?? agentAvatarUrl;
-  const handleToggle = React.useCallback(
-    (event: React.SyntheticEvent<HTMLDetailsElement>) => {
-      setIsExpanded(event.currentTarget.open);
-    },
-    [],
-  );
-
-  if (compactSummary.presentation === "message") {
-    return (
-      <div
-        className="not-prose w-full"
-        data-testid="transcript-tool-item"
-        title={timestampTitle}
-      >
-        <CompactMessageSummary
-          args={item.args}
-          avatarUrl={agentResolvedAvatarUrl}
-          description={buzzTool?.label}
-          displayName={agentLabel}
-          duration={duration}
-          hasArgs={hasArgs}
-          hasResult={hasResult}
-          isError={item.isError || item.status === "failed"}
-          label={compactSummary.label}
-          messageLink={messageLink}
-          preview={compactSummary.preview}
-          result={item.result}
-          timestamp={item.timestamp}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div
-      className="not-prose w-full"
-      data-testid="transcript-tool-item"
-      title={timestampTitle}
-    >
-      <details
-        className="group w-full"
-        onToggle={handleToggle}
-        open={isExpanded}
-      >
-        <summary
-          className={cn(
-            "max-w-full cursor-pointer list-none",
-            compactSummary.presentation === "inline" &&
-              "inline-flex items-center gap-1.5",
-            compactSummaryTone(),
-          )}
-        >
-          <CompactToolSummaryRow
-            action={compactSummary.action}
-            duration={duration}
-            fileEditSummary={compactSummary.fileEditSummary}
-            preview={compactSummary.preview}
-            thumbnailSrc={compactSummary.thumbnailSrc}
-            label={compactSummary.label}
-          />
-        </summary>
-
-        <ToolDetailBlocks
-          args={item.args}
-          description={buzzTool?.label}
-          fileEditDiff={compactSummary.fileEditDiff}
-          hasArgs={hasArgs}
-          hasResult={hasResult}
-          imagePreview={
-            compactSummary.thumbnailSrc != null && isExpanded
-              ? {
-                  src: compactSummary.thumbnailSrc,
-                  title: compactSummary.preview,
-                }
-              : null
-          }
-          isError={item.isError}
-          result={item.result}
-        />
-      </details>
-    </div>
-  );
-}
-
-function compactSummaryTone() {
-  return "text-muted-foreground/60 group-open:text-foreground";
-}
-
-function resolveImageSrc(source: string): string {
-  return isInlineImageData(source) ? source : rewriteRelayUrl(source);
-}
-
-function CompactToolSummaryRow({
-  action,
-  duration,
-  fileEditSummary,
-  label,
-  preview,
-  thumbnailSrc,
-}: {
-  action: AgentActivityAction | null;
-  duration: string | null;
-  fileEditSummary: CompactFileEditSummary | null;
-  label: string;
-  preview: string | null;
-  thumbnailSrc: string | null;
-}) {
-  const [thumbnailFailed, setThumbnailFailed] = React.useState(false);
-  const mutedTone = compactSummaryTone();
-  const resolvedThumbnail = React.useMemo(() => {
-    if (!thumbnailSrc || thumbnailFailed) return null;
-    return resolveImageSrc(thumbnailSrc);
-  }, [thumbnailFailed, thumbnailSrc]);
-  const actionLabel = fileEditSummary
-    ? null
-    : getCompactToolActionLabel(action, label, preview);
-
-  return (
-    <>
-      {fileEditSummary ? (
-        <CompactFileEditSummaryView summary={fileEditSummary} />
-      ) : actionLabel ? (
-        <ActivityRowLabel
-          object={actionLabel.object}
-          openToneScope="tool"
-          title={actionLabel.title}
-          verb={actionLabel.verb}
-        />
-      ) : (
-        <span className={cn("shrink-0 text-sm font-semibold", mutedTone)}>
-          {label}
-        </span>
-      )}
-      {!fileEditSummary && resolvedThumbnail ? (
-        <img
-          alt=""
-          className="h-5 w-auto max-w-12 shrink-0 rounded-sm object-cover"
-          decoding="async"
-          loading="lazy"
-          onError={() => setThumbnailFailed(true)}
-          src={resolvedThumbnail}
-          title={preview ?? undefined}
-        />
-      ) : !fileEditSummary && !actionLabel && preview ? (
-        <span
-          className={cn("min-w-0 max-w-48 truncate text-sm", mutedTone)}
-          title={preview}
-        >
-          {preview}
-        </span>
-      ) : null}
-      {duration ? (
-        <span className={cn("shrink-0 text-xs", mutedTone)}>{duration}</span>
-      ) : null}
-      <ChevronDown
-        className={cn(
-          "h-3.5 w-3.5 shrink-0 transition-transform group-open:rotate-180",
-          mutedTone,
-        )}
-      />
-    </>
-  );
-}
-
-function getCompactToolActionLabel(
-  action: AgentActivityAction | null,
-  label: string,
-  preview: string | null,
-): (ActivityRowLabelParts & { title?: string }) | null {
-  if (action) {
-    const object = action.object ?? preview ?? undefined;
-    return {
-      verb: action.verb,
-      object,
-      title: typeof object === "string" ? object : undefined,
-    };
-  }
-
-  const parts = splitActivityRowLabel(label);
-  if (!parts) return null;
-
-  if (!preview) return parts;
-
-  if (
-    label === "Ran command" ||
-    label === "Read file" ||
-    label === "Updated todos" ||
-    label === "Viewed image"
-  ) {
-    return { verb: parts.verb, object: preview, title: preview };
-  }
-
-  return parts;
-}
-
-function CompactFileEditSummaryView({
-  summary,
-}: {
-  summary: CompactFileEditSummary;
-}) {
-  return (
-    <ActivityRowLabel
-      className="max-w-72"
-      object={summary.filename}
-      openToneScope="tool"
-      stats={{
-        additions: summary.additions,
-        deletions: summary.deletions,
-      }}
-      title={summary.path}
-      verb="Edited"
-    />
-  );
-}
-
-function CompactMessageSummary({
-  args,
-  avatarUrl,
-  description,
-  displayName,
-  duration,
-  hasArgs,
-  hasResult,
-  isError,
-  label,
-  messageLink,
-  preview,
-  result,
-  timestamp,
-}: {
-  args: Record<string, unknown>;
-  avatarUrl: string | null;
-  description?: string;
-  displayName: string;
-  duration: string | null;
-  hasArgs: boolean;
-  hasResult: boolean;
-  isError: boolean;
-  label: string;
-  messageLink: SentMessageLink | null;
-  preview: string | null;
-  result: string;
-  timestamp: string;
-}) {
-  const [detailsOpen, setDetailsOpen] = React.useState(false);
-  const mutedTone = compactSummaryTone();
-  return (
-    <>
-      <div className="flex max-w-full flex-row items-start justify-start">
-        <UserAvatar
-          avatarUrl={avatarUrl}
-          className="mr-2 mt-1 shrink-0"
-          displayName={displayName}
-          size="xs"
-          testId="transcript-agent-sent-avatar"
-        />
-        <div className="flex max-w-[85%] min-w-0 flex-col items-start gap-1">
-          <div
-            className={cn(
-              "min-w-0 rounded-2xl border px-3 py-2 text-sm leading-relaxed shadow-sm",
-              isError
-                ? "border-destructive/25 bg-destructive/10 text-destructive"
-                : "border-primary/15 bg-primary/6 text-foreground",
-            )}
-            data-testid="transcript-tool-message-preview"
-          >
-            <p className="whitespace-pre-wrap wrap-break-word">
-              {preview || "Message content unavailable."}
-            </p>
-          </div>
-          <div className="inline-flex max-w-full items-center gap-1.5 px-1">
-            <TranscriptTimestamp
-              messageLink={messageLink}
-              timestamp={timestamp}
-            />
-            <button
-              aria-label="Show sent message context"
-              className={cn(
-                "inline-flex h-5 w-5 shrink-0 items-center justify-center rounded-sm transition-colors hover:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring",
-                mutedTone,
-              )}
-              data-testid="transcript-sent-message-context-button"
-              onClick={() => setDetailsOpen(true)}
-              title="Show sent message context"
-              type="button"
-            >
-              <CheckCheck className="h-3.5 w-3.5" />
-            </button>
-          </div>
-        </div>
-      </div>
-      <SentMessageContextDialog
-        args={args}
-        description={description}
-        duration={duration}
-        hasArgs={hasArgs}
-        hasResult={hasResult}
-        isError={isError}
-        label={label}
-        onOpenChange={setDetailsOpen}
-        open={detailsOpen}
-        preview={preview}
-        result={result}
-      />
-    </>
-  );
-}
-
-function SentMessageContextDialog({
-  args,
-  description,
-  duration,
-  hasArgs,
-  hasResult,
-  isError,
-  label,
-  onOpenChange,
-  open,
-  preview,
-  result,
-}: {
-  args: Record<string, unknown>;
-  description?: string;
-  duration: string | null;
-  hasArgs: boolean;
-  hasResult: boolean;
-  isError: boolean;
-  label: string;
-  onOpenChange: (open: boolean) => void;
-  open: boolean;
-  preview: string | null;
-  result: string;
-}) {
-  const sections = buildSentMessageContextSections({
-    args,
-    description,
-    hasArgs,
-    hasResult,
-    isError,
-    preview,
-    result,
-  });
-
-  return (
-    <Dialog onOpenChange={onOpenChange} open={open}>
-      <DialogContent className="max-w-xl overflow-hidden p-0">
-        <div className="flex max-h-[85vh] flex-col">
-          <DialogHeader className="px-6 pb-3 pt-5 pr-14">
-            <DialogTitle>Sent message context</DialogTitle>
-            <DialogDescription className="flex items-center gap-1.5">
-              <CheckCheck className="h-3.5 w-3.5 shrink-0" />
-              <span>{label}</span>
-              {duration ? <span className="shrink-0">{duration}</span> : null}
-            </DialogDescription>
-          </DialogHeader>
-          <div className="min-h-0 flex-1 overflow-y-auto px-6 pb-6 pt-2">
-            <SentMessageContextSections sections={sections} />
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-type SentMessageContextSection = {
-  body: string;
-  title: string;
-};
-
-function buildSentMessageContextSections({
-  args,
-  description,
-  hasArgs,
-  hasResult,
-  isError,
-  preview,
-  result,
-}: {
-  args: Record<string, unknown>;
-  description?: string;
-  hasArgs: boolean;
-  hasResult: boolean;
-  isError: boolean;
-  preview: string | null;
-  result: string;
-}): SentMessageContextSection[] {
-  const sections: SentMessageContextSection[] = [];
-  if (preview) {
-    sections.push({ title: "Message", body: preview });
-  }
-  if (description) {
-    sections.push({ title: "Tool", body: description });
-  }
-  if (hasArgs) {
-    sections.push({
-      title: "Parameters",
-      body: JSON.stringify(args, null, 2),
-    });
-  }
-  if (hasResult) {
-    sections.push({
-      title: isError ? "Error" : "Result",
-      body: formatCodeValue(result),
-    });
-  }
-  if (sections.length === 0) {
-    sections.push({
-      title: "Status",
-      body: "Waiting for tool details.",
-    });
-  }
-  return sections;
-}
-
-function SentMessageContextSections({
-  sections,
-}: {
-  sections: SentMessageContextSection[];
-}) {
-  return (
-    <div className="space-y-3" data-testid="transcript-sent-message-context">
-      {sections.map((section) => (
-        <SentMessageContextSectionAccordion
-          key={`${section.title}:${section.body.slice(0, 48)}`}
-          section={section}
-        />
-      ))}
-    </div>
-  );
-}
-
-function SentMessageContextSectionAccordion({
-  section,
-}: {
-  section: SentMessageContextSection;
-}) {
-  const [open, setOpen] = React.useState(false);
-  const body = section.body.trim();
-
-  return (
-    <article className="overflow-hidden rounded-2xl bg-muted/40">
-      <button
-        aria-expanded={open}
-        className="w-full px-4 py-3 text-left transition-colors hover:bg-muted/50"
-        onClick={() => setOpen((value) => !value)}
-        type="button"
-      >
-        <div className="flex items-start gap-3">
-          <div className="min-w-0 flex-1">
-            <div
-              className={cn(
-                "text-sm font-semibold text-foreground",
-                !open && "line-clamp-2",
-              )}
-            >
-              {section.title}
-            </div>
-            <div
-              className={cn(
-                "mt-1 whitespace-pre-wrap break-all text-xs leading-5 text-foreground/70",
-                !open && "line-clamp-2",
-              )}
-            >
-              {body.length > 0 ? (
-                body
-              ) : (
-                <span className="italic text-foreground/50">No metadata.</span>
-              )}
-            </div>
-          </div>
-          <ChevronDown
-            className={cn(
-              "mt-0.5 h-4 w-4 shrink-0 text-muted-foreground transition-transform",
-              open && "rotate-180",
-            )}
-          />
-        </div>
-      </button>
-    </article>
-  );
-}
-
-type SentMessageLink = {
-  channelId: string;
-  messageId: string;
-};
-
-function getSentMessageLink(
-  item: Extract<TranscriptItem, { type: "tool" }>,
-): SentMessageLink | null {
-  if (item.status !== "completed" || item.isError) {
-    return null;
-  }
-
-  if (item.descriptor?.renderClass !== "message") {
-    return null;
-  }
-
-  const channelId =
-    item.channelId ?? getToolString(item.args, ["channel_id", "channelId"]);
-  if (!channelId) {
-    return null;
-  }
-
-  const resultRecord = getMessageSendResultRecord(item.result);
-  if (!resultRecord || resultRecord.accepted === false) {
-    return null;
-  }
-
-  const messageId = getToolString(resultRecord, [
-    "event_id",
-    "eventId",
-    "message_id",
-    "messageId",
-  ]);
-  if (!messageId) {
-    return null;
-  }
-
-  return {
-    channelId,
-    messageId,
-  };
-}
-
-function getMessageSendResultRecord(
-  result: string,
-): Record<string, unknown> | null {
-  const parsed = parseToolResultValue(result);
-  const directRecord = asRecord(parsed);
-  if (getMessageEventId(directRecord)) {
-    return directRecord;
-  }
-
-  const stdout = getToolString(directRecord, ["stdout"]);
-  if (!stdout) {
-    return null;
-  }
-
-  const stdoutRecord = asRecord(parseToolResultValue(stdout));
-  return getMessageEventId(stdoutRecord) ? stdoutRecord : null;
-}
-
-function getMessageEventId(record: Record<string, unknown>) {
-  return getToolString(record, [
-    "event_id",
-    "eventId",
-    "message_id",
-    "messageId",
-  ]);
-}
-
-function ViewImageToolPreview({
-  src,
-  title,
-}: {
-  src: string;
-  title: string | null;
-}) {
-  const [lightboxOpen, setLightboxOpen] = React.useState(false);
-  const [imageFailed, setImageFailed] = React.useState(false);
-  const resolvedSrc = React.useMemo(() => resolveImageSrc(src), [src]);
-  const alt = title ?? "Viewed image";
-
-  if (imageFailed) {
-    return null;
-  }
-
-  return (
-    <>
-      {/* biome-ignore lint/a11y/useKeyWithClickEvents: opens lightbox on click */}
-      <img
-        alt={alt}
-        className="ml-1.5 block max-h-64 max-w-[min(24rem,calc(100%-0.375rem))] cursor-pointer rounded-lg object-contain"
-        decoding="async"
-        loading="lazy"
-        onClick={() => setLightboxOpen(true)}
-        onError={() => setImageFailed(true)}
-        src={resolvedSrc}
-        title={title ?? undefined}
-      />
-      <ImageLightbox
-        alt={alt}
-        onOpenChange={setLightboxOpen}
-        open={lightboxOpen}
-        src={resolvedSrc}
-      />
-    </>
-  );
-}
-
-function ImageLightbox({
-  alt,
-  onOpenChange,
-  open,
-  src,
-}: {
-  alt: string;
-  onOpenChange: (open: boolean) => void;
-  open: boolean;
-  src: string;
-}) {
-  return (
-    <DialogPrimitive.Root open={open} onOpenChange={onOpenChange}>
-      <DialogPrimitive.Portal>
-        <DialogPrimitive.Overlay className="fixed inset-0 z-50 bg-black/80 data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0" />
-        <DialogPrimitive.Content
-          className="fixed inset-0 z-50 flex items-center justify-center p-8"
-          onInteractOutside={(event) => event.preventDefault()}
-          onPointerDownOutside={(event) => event.preventDefault()}
-        >
-          <DialogPrimitive.Title className="sr-only">
-            {alt}
-          </DialogPrimitive.Title>
-          <DialogPrimitive.Description className="sr-only">
-            Full-size image preview. Press Escape or click outside the image to
-            close.
-          </DialogPrimitive.Description>
-          <DialogPrimitive.Close
-            aria-label="Close lightbox"
-            className="absolute inset-0 cursor-default"
-          />
-          <img
-            alt={alt}
-            className="relative max-h-[90vh] max-w-[90vw] rounded-lg object-contain"
-            src={src}
-          />
-          <DialogPrimitive.Close className="absolute right-4 top-4 rounded-full bg-black/50 p-2 text-white/80 transition-colors hover:bg-black/70 hover:text-white focus:outline-hidden focus:ring-2 focus:ring-white/30">
-            <svg
-              aria-hidden="true"
-              fill="none"
-              height="20"
-              viewBox="0 0 24 24"
-              width="20"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M18 6L6 18M6 6l12 12"
-                stroke="currentColor"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth="2"
-              />
-            </svg>
-          </DialogPrimitive.Close>
-        </DialogPrimitive.Content>
-      </DialogPrimitive.Portal>
-    </DialogPrimitive.Root>
-  );
-}
-
-function ToolDetailBlocks({
-  args,
-  description,
-  fileEditDiff,
-  hasArgs,
-  hasResult,
-  imagePreview,
-  isError,
-  result,
-}: {
-  args: Record<string, unknown>;
-  description?: string;
-  fileEditDiff: FileEditDiff | null;
-  hasArgs: boolean;
-  hasResult: boolean;
-  imagePreview: { src: string | null; title: string | null } | null;
-  isError: boolean;
-  result: string;
-}) {
-  const showFileEditDiff =
-    fileEditDiff && hasFileEditLineDiff(fileEditDiff) && !isError;
-  const showParameters = hasArgs && !showFileEditDiff;
-
-  return (
-    <div className="space-y-4 py-2 text-popover-foreground outline-hidden">
-      {description ? (
-        <p className="max-w-2xl text-xs leading-5 text-muted-foreground">
-          {description}
-        </p>
-      ) : null}
-      {imagePreview?.src ? (
-        <ViewImageToolPreview
-          src={imagePreview.src}
-          title={imagePreview.title}
-        />
-      ) : null}
-      {showParameters ? (
-        <ToolCodeBlock
-          label="Parameters"
-          tone="muted"
-          value={JSON.stringify(args, null, 2)}
-        />
-      ) : null}
-      {hasResult ? (
-        showFileEditDiff ? (
-          <FileEditDiffBlock diff={fileEditDiff} />
-        ) : (
-          <ToolCodeBlock
-            label={isError ? "Error" : "Result"}
-            tone={isError ? "error" : "muted"}
-            value={result}
-          />
-        )
-      ) : null}
-      {!showParameters && !hasResult ? (
-        <p className="text-sm text-muted-foreground/80">
-          Waiting for tool details.
-        </p>
-      ) : null}
-    </div>
-  );
-}
-
-function ToolCodeBlock({
-  label,
-  tone,
-  value,
-}: {
-  label: string;
-  tone: "muted" | "error";
-  value: string;
-}) {
-  return (
-    <div className="space-y-2 overflow-hidden">
-      <h4 className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-        {label}
-      </h4>
-      <pre
-        className={cn(
-          "max-h-64 overflow-auto whitespace-pre-wrap wrap-break-word rounded-md px-3 py-2 font-mono text-xs leading-5",
-          tone === "error"
-            ? "bg-destructive/10 text-destructive"
-            : "bg-muted/50 text-foreground",
-        )}
-      >
-        {formatCodeValue(value)}
-      </pre>
-    </div>
-  );
-}
+export { ToolItem } from "./AgentSessionToolItem/ToolItem";
